Refactoring components folder
