+++ upnext-web-app/src/UpNext/components/QueuePages/ListGroupSelect.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { useEffect, useState } from "react";
 import { Form } from "react-bootstrap";
-import * as groupClient from "../clients/groupClient";
+import * as groupClient from "../../clients/groupClient";
 import { useSelector } from "react-redux";
 
 export default function ListGroupSelect({
