+++ upnext-web-app/src/UpNext/components/styles/MediaSearch.css
