+++ upnext-web-app/src/UpNext/components/styles/QueueList.css
